refactor: define header for detail-panel.tsx
