+++ src/views/app/detail-panel/conversation-mode/conversation-message-preview-wrapper.tsx
