+++ src/views/app/detail-panel/conversation-mode/conversation-preview-panel.tsx
