+++ src/views/app/detail-panel/eml-preview-panel-container.tsx
@@ -8,7 +8,7 @@ import React, { useEffect, useState } from 'react';
 
 import { useParams } from 'react-router-dom';
 
-import { MessagePreviewPanel } from './messages/message-preview-panel';
+import { MessagePreviewPanel } from './message-mode/message-preview-panel';
 import type { EmlRouteParams } from '../../../types/routes';
 import { getMsgSoapApi } from 'api/get-msg-soap-api';
 import { isFocusModeMailView } from 'helpers/external-tabs';
