+++ src/views/app/detail-panel/message-mode/message-preview-panel-container.tsx
